backend over

+++ Backend/controllers/goalscontroller.js
@@ -1,7 +1,8 @@
 const asyncHandler=require('express-async-handler');
 const Goal=require('../Model/goalModel');
+const User=require('../Model/userModel');
 const getGoals=asyncHandler(async(req,res)=>{
-    const goals=await Goal.find();
+    const goals=await Goal.find({user:req.user.id});
     res.status(200).json(goals);
 });
 const setGoals=asyncHandler(async(req,res)=>{
@@ -14,14 +15,25 @@ const setGoals=asyncHandler(async(req,res)=>{
     }
     const goal=await Goal.create({
         text:req.body.text,
+        user:req.user.id
     })
     res.status(201).json(goal);
 });
 const putGoals=asyncHandler(async(req,res)=>{
+
     const goal=await Goal.findById(req.params.id);
     if(!goal){
         res.status(400);
         throw new Error('goal not found')
+    }
+    const user=await User.findById(req.user.id);
+    if(!user){
+        res.status(401);
+        throw new Error('user not found');
+    }
+    if(goal.user.toString() !=user.id){
+        res.status(401);
+        throw new Error('user not authorized');
     }
     const updatedGoal=await Goal.findByIdAndUpdate(req.params.id,req.body,{new:true})
     res.status(200).json(updatedGoal);
@@ -31,6 +43,15 @@ const deleteGoals=asyncHandler(async(req,res)=>{
     if(!goal){
         res.status(400);
         throw new Error('goal not found')
+    }
+    const user=await User.findById(req.user.id);
+    if(!user){
+        res.status(401);
+        throw new Error('user not found');
+    }
+    if(goal.user.toString() !=user.id){
+        res.status(401);
+        throw new Error('user not authorized');
     }
     await Goal.deleteOne({ _id: req.params.id });
     //(or)
